requirements.txt
kserve==0.10.2
pydantic==1.10.12
requests==2.29.0
aiohttp==3.8.3
aiofiles==23.2.1
schema
pyyaml
click
pytest
mmh3
boto3